Go through the code in `mill.main.maven.BuildGen`, try getting the necessary build information with the Gradle Tooling API, and extract a `CommonBuildGen` trait
There are some issues in the current direction of getting build information using the Gradle Tooling API:

- There isn't a built-in way to get the javac options.
- The `ProjectPublications` and `GradlePublication` models don't seem to provide enough information, such as license, the version control URL, and the developer.

Therefore, I am considering an alternative approach: retrieve the necessary information in an added task in a Gradle init script and pass it to the Mill task. To make the code more modular and maintainable, I can implement a Gradle plugin in a separate project and apply it in the init script, like how it's done in https://github.com/microsoft/build-server-for-gradle.

There is another alternative approach: run the build server and communicate with it directly with support from https://github.com/microsoft/build-server-for-gradle. However, there are also some issues:

- The project is not published to Maven Central, so it may take some effort to set it up and run.
- There is no BSP request to retrieve publish metadata.

There is one last approach: enable all Gradle projects to publish to Maven Local, possibly with the Gradle init script too, and then generate with `mill.main.maven.BuildGen`. Maven doesn't distinguish inter-subproject dependencies and third-party libraries, so the inter-subproject dependencies should be able to be recovered with the existing Maven `BuildGen`.